Cast Average() to long
Cast mean so it compiles.

+++ src/AdventOfCode/Puzzles/Y2021/Day07.cs
@@ -34,7 +34,7 @@ public sealed class Day07 : Puzzle
     {
         if (withVariableBurnRate)
         {
-            return submarines.Average();
+            return (long)submarines.Average();
         }
 
         submarines = submarines.OrderBy((p) => p).ToList();
